+++ src/interpreter/Interpreter.c
@@ -72,7 +72,7 @@ static void pop_frame_and_push_result(pVMObject result) {
     // of arguments
     pVMFrame prev_frame = pop_frame();
     pVMMethod method = SEND(prev_frame, get_method);
-    int number_of_arguments = SEND(method, get_number_of_arguments);
+    int64_t number_of_arguments = SEND(method, get_number_of_arguments);
         
     // Pop the arguments
     for(int i = 0; i < number_of_arguments; i++)
@@ -125,7 +125,7 @@ static void do_dup(void) {
 }
 
 
-static void do_push_local(int bytecode_index) {
+static void do_push_local(size_t bytecode_index) {
     // Handle the push local bytecode
     pVMMethod method = _METHOD;
     uint8_t bc1 = SEND(method, get_bytecode, bytecode_index + 1);
@@ -137,7 +137,7 @@ static void do_push_local(int bytecode_index) {
 }
 
 
-static void do_push_argument(int bytecode_index) {
+static void do_push_argument(size_t bytecode_index) {
     // Handle the push argument bytecode
     pVMMethod method = _METHOD;
     uint8_t bc1 = SEND(method, get_bytecode, bytecode_index + 1);
@@ -149,7 +149,7 @@ static void do_push_argument(int bytecode_index) {
 }
 
 
-static void do_push_field(int bytecode_index) {
+static void do_push_field(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the push field bytecode
     pVMSymbol field_name =
@@ -157,7 +157,7 @@ static void do_push_field(int bytecode_index) {
     
     // Get the field index from the field name
     pVMObject self = _SELF;
-    int field_index = SEND(self, get_field_index, field_name);
+    int64_t field_index = SEND(self, get_field_index, field_name);
     
     pVMObject o = SEND(self, get_field, field_index);
     // Push the field with the computed index onto the stack
@@ -165,13 +165,13 @@ static void do_push_field(int bytecode_index) {
 }
 
 
-static void do_push_block(int bytecode_index) {
+static void do_push_block(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the push block bytecode
     pVMMethod block_method = (pVMMethod)SEND(method, 
                                               get_constant, bytecode_index);
         
-    int number_of_arguments = SEND(block_method, get_number_of_arguments);
+    int64_t number_of_arguments = SEND(block_method, get_number_of_arguments);
     // Push a new block with the current get_frame() as context onto the stack
     SEND(_FRAME, push,
          (pVMObject) Universe_new_block(block_method,
@@ -180,7 +180,7 @@ static void do_push_block(int bytecode_index) {
 }
 
 
-static void do_push_constant(int bytecode_index) {
+static void do_push_constant(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     
     // Handle the push constant bytecode
@@ -189,7 +189,7 @@ static void do_push_constant(int bytecode_index) {
 }
 
 
-static void do_push_global(int bytecode_index) {
+static void do_push_global(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the push global bytecode
     pVMSymbol global_name = (pVMSymbol)SEND(method,
@@ -216,7 +216,7 @@ static void do_pop(void) {
 }
 
 
-static void do_pop_local(int bytecode_index) {
+static void do_pop_local(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the pop local bytecode
     uint8_t bc1 = SEND(method, get_bytecode, bytecode_index + 1);
@@ -228,7 +228,7 @@ static void do_pop_local(int bytecode_index) {
 }
 
 
-static void do_pop_argument(int bytecode_index) {
+static void do_pop_argument(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the pop argument bytecode
     uint8_t bc1 = SEND(method, get_bytecode, bytecode_index + 1);
@@ -239,7 +239,7 @@ static void do_pop_argument(int bytecode_index) {
 }
 
 
-static void do_pop_field(int bytecode_index) {
+static void do_pop_field(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the pop field bytecode
     pVMSymbol field_name = (pVMSymbol)SEND(method, 
@@ -247,7 +247,7 @@ static void do_pop_field(int bytecode_index) {
                 
     // Get the field index from the field name
     pVMObject self = _SELF;
-    int field_index = SEND(self, get_field_index, field_name);
+    int64_t field_index = SEND(self, get_field_index, field_name);
     
     // Set the field with the computed index to the value popped from the stack
     pVMObject o = SEND(_FRAME, pop);
@@ -255,7 +255,7 @@ static void do_pop_field(int bytecode_index) {
 }
 
 
-static void do_send(int bytecode_index) {
+static void do_send(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the send bytecode
     pVMSymbol signature = (pVMSymbol)SEND(method, 
@@ -273,7 +273,7 @@ static void do_send(int bytecode_index) {
 }
 
 
-static void do_super_send(int bytecode_index) {
+static void do_super_send(size_t bytecode_index) {
     pVMMethod method = _METHOD;
     // Handle the super send bytecode
     pVMSymbol signature = (pVMSymbol)SEND(method, 
@@ -374,7 +374,7 @@ void Interpreter_start(void) {
     // iterate over the bytecodes
     while(true) {
         // get the current bytecode index
-        int bytecode_index = SEND(_FRAME, get_bytecode_index);
+        size_t bytecode_index = SEND(_FRAME, get_bytecode_index);
         // get the current bytecode
         pVMMethod method = Interpreter_get_method();
         uint8_t bytecode = SEND(method, get_bytecode, bytecode_index);
@@ -386,7 +386,7 @@ void Interpreter_start(void) {
             Disassembler_dump_bytecode(_FRAME, method, bytecode_index);
         
         // compute the next bytecode index
-        int next_bytecode_index = bytecode_index + bytecode_length;
+        size_t next_bytecode_index = bytecode_index + bytecode_length;
         // update the bytecode index of the frame
         SEND(_FRAME, set_bytecode_index, next_bytecode_index);
